fix: backup apps/theme as the first step
